requirements.txt
# requirements.txt
# move-agent-kit>=1.0.0  # Commented out as it's only available for JS/TS
flask>=2.0.0
flask-restful>=0.3.9
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.27.1
pillow>=9.0.0
websockets>=10.0
aptos-sdk>=0.5.1
tabulate>=0.9.0
